+++ test/bigtest/index.js
@@ -1,7 +1,0 @@
-import 'core-js/stable';
-import 'regenerator-runtime/runtime';
-
-// require all modules ending in "-test" from the current directory and
-// all subdirectories
-const requireTest = require.context('./tests/', true, /-test/);
-requireTest.keys().forEach(requireTest);