+++ test/bigtest/network/models/loan.js
@@ -1,6 +1,0 @@
-import { Model, belongsTo } from 'miragejs';
-
-export default Model.extend({
-  item: belongsTo(),
-  user: belongsTo()
-});