+++ test/bigtest/network/scenarios/requestQueueScenario.js
@@ -1,3 +1,0 @@
-export default function (server) {
-  server.createList('request', 3, { requestType: 'Hold' }, 'withPagedItems', 'withCallNumber');
-}