+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * UI tests replacement with RTL/Jest for src/MoveRequestManager.js. Refs UIREQ-877.
 * Increase code coverage for Jest/RTL tests. Refs UIREQ-995.
 * Convert primary search results view to use Prev/Next pagination. Refs UIREQ-965.
+* Remove bigtest tests and associated dependencies. Refs UIREQ-999.
 
 ## [8.0.2](https://github.com/folio-org/ui-requests/tree/v8.0.2) (2023-03-29)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v8.0.1...v8.0.2)